Refactor audio service into backend and service modules
